backend/requirements.txt
flask==1.1.1
gunicorn==20.0.4
flask-mongoengine
tensorflow==2.3
flask_mongoengine
flask_cors
pillow
dnspython